Ground truth (#4)
Project restructure and ground truthing and utils modules
