+++ .gitignore
@@ -1,0 +1,19 @@
+# Pycharm
+.idea
+
+# venv
+.venv*
+
+# python
+**/__pycache__
+*.pyc
+
+# pyenv
+.python-version
+
+# .egg-info
+*.egg-info/
+
+# config files
+_DENA/config/*
+!_DENA/config/template.config
