+++ _DENA/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+_DENA_DIR = os.path.dirname(os.path.abspath(__file__))
