+++ _DENA/config/template.config
@@ -1,0 +1,14 @@
+[database:overflights]
+name =
+username =
+password =
+port =
+host =
+
+[data]
+site_metadata =
+nvspl_archive =
+adsb =
+
+[project]
+dir =
