+++ _DENA/requirements.txt
@@ -1,0 +1,48 @@
+affine==2.3.1
+attrs==22.1.0
+certifi==2022.6.15
+charset-normalizer==2.1.0
+click==8.1.3
+click-plugins==1.1.1
+cligj==0.7.2
+colorama==0.4.5
+contextily==1.2.0
+cycler==0.11.0
+Fiona==1.8.21
+fonttools==4.34.4
+future==0.18.2
+GDAL==3.4.2
+geographiclib==1.52
+geopandas==0.10.2
+geopy==2.2.0
+greenlet==1.1.2
+idna==3.3
+importlib-metadata==4.12.0
+iyore @ git+https://github.com/nationalparkservice/iyore.git
+joblib==1.1.0
+kiwisolver==1.4.4
+matplotlib==3.5.2
+mercantile==1.2.1
+munch==2.5.0
+numpy==1.21.6
+packaging==21.3
+pandas==1.1.5
+Pillow==9.1.1
+psycopg2==2.9.3
+pygeos==0.12.0
+pyparsing==3.0.9
+pyproj==3.2.1
+python-dateutil==2.8.2
+pytz==2022.1
+rasterio==1.2.10
+requests==2.28.1
+scipy==1.7.3
+Shapely==1.8.2
+six==1.16.0
+snuggs==1.4.7
+SQLAlchemy==1.4.39
+tqdm==4.64.0
+typing_extensions==4.3.0
+urllib3==1.26.11
+xyzservices==2022.6.0
+zipp==3.8.1
