+++ legacy_code/NestablePool.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+"""
+This is taken from StackOverflow:
+https://stackoverflow.com/questions/6974695/python-process-pool-non-daemonic
+
+All credit for the nestable pool to Massimiliano
+"""
+
+import multiprocessing
+# We must import this explicitly, it is not imported by the top-level
+# multiprocessing module.
+import multiprocessing.pool
+
+class NoDaemonProcess(multiprocessing.Process):
+    @property
+    def daemon(self):
+        return False
+
+    @daemon.setter
+    def daemon(self, value):
+        pass
+
+
+class NoDaemonContext(type(multiprocessing.get_context())):
+    Process = NoDaemonProcess
+
+# We sub-class multiprocessing.pool.Pool instead of multiprocessing.Pool
+# because the latter is only a wrapper function, not a proper class.
+class NestablePool(multiprocessing.pool.Pool):
+    def __init__(self, *args, **kwargs):
+        kwargs['context'] = NoDaemonContext()
+        super(NestablePool, self).__init__(*args, **kwargs)
