+++ nps_active_space/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+_ACTIVE_SPACE_DIR = os.path.dirname(os.path.abspath(__file__))
